deck: layout is a deck helper, and a pair's unresolved forms are its vocabulary
layout.ts sat loose at the top of src doing the same kind of work as everything
in util, so it moves there. navigation-target held two functions about a pair
before it resolves, which is what navigation is already about, so they merge.

util/ is the deck's pure helpers and url/ is the machinery that reads and writes
the address bar, which its barrel now says. Everything util exports carries a
line saying what it is.
