+++ packages/plugins/plugin-deck/src/util/index.ts
@@ -3,6 +3,7 @@
 //
 
 export * from './companion-anchor';
+export * from './layout';
 export * from './layoutAppliesTopbar';
 export * from './migrate-persisted-state';
 export * from './resolve-sidebar-state';
